requirements.txt
Django==1.6
Markdown==2.4.1
South==1.0
argparse==1.2.1
djangorestframework==2.3.14
wsgiref==0.1.2
